# [LCS] uleb128 for lengths and variant indexes