--- a/network/src/protocols/wire/messaging/v1/test.rs
+++ b/network/src/protocols/wire/messaging/v1/test.rs
@@ -14,7 +14,7 @@ fn protocol_id_serialization() -> lcs::Result<()> {
 #[test]
 fn error_code() -> lcs::Result<()> {
     let error_code = ErrorCode::TimedOut;
-    assert_eq!(lcs::to_bytes(&error_code)?, vec![0x01, 0, 0, 0]);
+    assert_eq!(lcs::to_bytes(&error_code)?, vec![0x01]);
     Ok(())
 }
 
@@ -31,9 +31,9 @@ fn rpc_request() -> lcs::Result<()> {
         // [25, 0, 0, 0] -> request_id
         // [0] -> protocol_idx
         // [0] -> priority
-        // [4, 0, 0, 0] -> length of raw_request
+        // [4] -> length of raw_request
         // [0, 1, 2, 3] -> raw_request bytes
-        vec![25, 0, 0, 0, 0, 0, 4, 0, 0, 0, 0, 1, 2, 3]
+        vec![25, 0, 0, 0, 0, 0, 4, 0, 1, 2, 3]
     );
     Ok(())
 }