… language/stdlib/staged/stdlib.mv
… language/tools/vm-genesis/genesis/genesis.blob
… types/src/unit_tests/canonical_serialization_examples.rs