## Motivation

Most vectors have a short lengths and most variant indexes do not exceed 127.
By using the ULEB128 encoding to represent these numbers, we reduce the data overhead to 1 byte instead of 4 bytes.

## Test Plan
```
cargo x test -p libra-canonical-serialization
```