@vgao1996 2020-03-17
Thanks! This addresses our most immediate concern! There's still some unclarity around the IDL (or source of truth), but we will think and talk about them.