@matbd 2020-03-19
With Victor and Runtian's help, I figured I should run `(cd language/stdlib; cargo run) && (cd language/tools/vm-genesis; cargo run)` to regenerate some staged artifacts.