@dariorussi 2020-03-21
at some point it would be nice if we had an API from LCS that allows us to call into it for ULEB and other de/serialization