@bmwill 2020-03-21
@dariorussi I think we'd need to be careful with what internals get exposed, I'd like to avoid getting into the situation we had before where we were hand implementing ser/de resulting in types violating the spec.