@davidiw 2020-03-21
> @dariorussi I think we'd need to be careful with what internals get exposed, I'd like to avoid getting into the situation we had before where we were hand implementing ser/de resulting in types violating the spec.

I think some of the work @matbd is doing on creating an IDL that ideally can be derived from (maybe into) Rust code and at least derived from Move code will make it easier to implement the appropriate ability to parse it and enforce a specification. At which point, the risk of exposing internals would be a non-issue. Though arguably, we have a lot of non-trivial work to do to get there.