@matbd 2020-03-24
@davidiw I ran the command below with different parameters before and after this commit and found that we are decreasing the size of the DB between 0.8% and 1.2%.

```
cargo run --release -p executor-benchmark -- --num-transfer-blocks 20 --num-accounts 5000 --db-dir $PWD/test
```