@davidiw 2020-03-25
Execution performance is all over the place but here are some samples. Data reduction is clearly down by roughly 3.9%. Note that at smaller loads the number was closer to 1% (as @matbd noted above). I'd be curious what the long term steady state win is... but I don't think we have the tooling to figure that out.

Will follow up with a final review of the PR.

Before: 10917 MB
INFO 2020-03-25 02:52:43 execution/executor-benchmark/src/lib.rs:262 Version: 1498000. execute time: 591 ms. commit time: 180 ms. TPS: 647.
INFO 2020-03-25 02:52:44 execution/executor-benchmark/src/lib.rs:262 Version: 1498500. execute time: 624 ms. commit time: 173 ms. TPS: 626.
INFO 2020-03-25 02:52:45 execution/executor-benchmark/src/lib.rs:262 Version: 1499000. execute time: 648 ms. commit time: 190 ms. TPS: 595.
INFO 2020-03-25 02:52:45 execution/executor-benchmark/src/lib.rs:262 Version: 1499500. execute time: 596 ms. commit time: 200 ms. TPS: 627.
INFO 2020-03-25 02:52:46 execution/executor-benchmark/src/lib.rs:262 Version: 1500000. execute time: 565 ms. commit time: 209 ms. TPS: 644.
After: 10488 MB
INFO 2020-03-25 01:55:14 execution/executor-benchmark/src/lib.rs:262 Version: 1498000. execute time: 574 ms. commit time: 183 ms. TPS: 659.
INFO 2020-03-25 01:55:14 execution/executor-benchmark/src/lib.rs:262 Version: 1498500. execute time: 553 ms. commit time: 191 ms. TPS: 671.
INFO 2020-03-25 01:55:15 execution/executor-benchmark/src/lib.rs:262 Version: 1499000. execute time: 559 ms. commit time: 182 ms. TPS: 673.
INFO 2020-03-25 01:55:16 execution/executor-benchmark/src/lib.rs:262 Version: 1499500. execute time: 550 ms. commit time: 179 ms. TPS: 685.
INFO 2020-03-25 01:55:17 execution/executor-benchmark/src/lib.rs:262 Version: 1500000. execute time: 574 ms. commit time: 197 ms. TPS: 647.