@matbd 2020-03-25
rebased (solving conflicts) + improved docs